monitoring_agents/requirements.txt
# Core dependencies
aiohttp>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0
python-dateutil>=2.8.0

# AI dependencies
google-generativeai>=0.3.0

# Technical analysis (optional)
pandas-ta>=0.3.14b
finta>=1.3

# Networking and APIs
requests>=2.28.0
websockets>=11.0

# Caching and storage
redis>=4.5.0

# Configuration
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
